Refactor out operator commands
