+++ docs/reference/jsctl_operator_deploy.md
@@ -33,5 +33,5 @@ jsctl operator deploy [flags]
 
 ### SEE ALSO
 
-* [jsctl operator](jsctl_operator.md)	 - Subcommands for managing the jetstack operator
+* [jsctl operator](jsctl_operator.md)	 - Subcommands for managing the Jetstack operator
